Separar interfaces para cumplir con el principio de interfaz de segregación (ISP)
