+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/application/ports/in/GetAndSaveMovieInfoUseCase.java
@@ -2,7 +2,6 @@ package com.peliculas.peliculasapp.application.ports.in;
 
 import com.peliculas.peliculasapp.dto.MovieDTO;
 
-public interface GetAndSaveInfoUseCase {
+public interface GetAndSaveMovieInfoUseCase {
     MovieDTO getAndSaveMovieInfo(long movieId);
-    void getAndSaveTvSeriesInfo(long tvSeriesId);
 }
