+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/application/ports/in/GetAndSaveTvSeriesInfoUseCase.java
@@ -1,0 +1,5 @@
+package com.peliculas.peliculasapp.application.ports.in;
+
+public interface GetAndSaveTvSeriesInfoUseCase {
+    void getAndSaveTvSeriesInfo(long tvSeriesId);
+}
