+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/application/services/MovieService.java
@@ -1,14 +1,14 @@
 package com.peliculas.peliculasapp.application.services;
-import com.peliculas.peliculasapp.application.ports.in.GetAndSaveInfoUseCase;
+import com.peliculas.peliculasapp.application.ports.in.GetAndSaveMovieInfoUseCase;
 import com.peliculas.peliculasapp.dto.MovieDTO;
 import org.springframework.stereotype.Service;
 
 
 @Service
 public class MovieService {
-    private final GetAndSaveInfoUseCase getAndSaveInfoUseCase;
+    private final GetAndSaveMovieInfoUseCase getAndSaveInfoUseCase;
 
-    public MovieService(GetAndSaveInfoUseCase getAndSaveInfoUseCase) {
+    public MovieService(GetAndSaveMovieInfoUseCase getAndSaveInfoUseCase) {
         this.getAndSaveInfoUseCase = getAndSaveInfoUseCase;
     }
