+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/application/services/TvSeriesService.java
@@ -1,0 +1,8 @@
+package com.peliculas.peliculasapp.application.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class TvSeriesService {
+    //TODO: implement
+}
