+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/application/usecases/GetAndSaveTvSeriesInfoUseCaseImpl.java
@@ -1,0 +1,11 @@
+package com.peliculas.peliculasapp.application.usecases;
+import com.peliculas.peliculasapp.application.ports.in.GetAndSaveTvSeriesInfoUseCase;
+import org.springframework.stereotype.Service;
+
+@Service
+public class GetAndSaveTvSeriesInfoUseCaseImpl implements GetAndSaveTvSeriesInfoUseCase {
+    @Override
+    public void getAndSaveTvSeriesInfo(long tvSeriesId) {
+        // TODO: implement
+    }
+}
