+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/dto/TvSeriesDTO.java
@@ -1,0 +1,5 @@
+package com.peliculas.peliculasapp.dto;
+
+public class TvSeriesDTO {
+    // TODO: Implement
+}
