+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/infrastructure/controllers/TvSeriesController.java
@@ -1,23 +1,10 @@
 package com.peliculas.peliculasapp.infrastructure.controllers;
-import com.peliculas.peliculasapp.application.ports.in.GetAndSaveInfoUseCase;
+import com.peliculas.peliculasapp.application.ports.in.GetAndSaveMovieInfoUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/tvseries")
 public class TvSeriesController {
-    private final GetAndSaveInfoUseCase getAndSaveInfoUseCase;
-
-    @Autowired
-    public TvSeriesController(GetAndSaveInfoUseCase getAndSaveInfoUseCase) {
-        this.getAndSaveInfoUseCase = getAndSaveInfoUseCase;
-    }
-
-    @GetMapping("/{id}")
-    public void getAndSaveTvSeriesInfo(@PathVariable long id) {
-        getAndSaveInfoUseCase.getAndSaveTvSeriesInfo(id);
-    }
+    // TODO: Implement
 }
