+++ peliculas-app/src/main/java/com/peliculas/peliculasapp/infrastructure/entities/MovieEntity.java
@@ -24,6 +24,7 @@ public class MovieEntity {
     private List<ProductionCompaniesEntity> productionCompanies;
 
     @OneToMany(cascade = CascadeType.ALL)
+    @JoinColumn(name = "movie_id", referencedColumnName = "id")
     private List<GenreEntity> genres;
 
     @OneToMany
